refactor; rename build.py to build_cmd.py
