+++ cognite_toolkit/_cdf_tk/commands/__init__.py
@@ -1,7 +1,7 @@
 from ._populate import PopulateCommand
 from ._purge import PurgeCommand
 from .auth import AuthCommand
-from .build import BuildCommand
+from .build_cmd import BuildCommand
 from .clean import CleanCommand
 from .collect import CollectCommand
 from .deploy import DeployCommand
